+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 7.1.19
+
+### Patch Changes
+
+- Updated dependencies [9aad939]
+    - @khanacademy/wonder-blocks-tokens@12.0.1
+
 ## 7.1.18
 
 ### Patch Changes