+++ packages/wonder-blocks-grid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-grid
 
+## 3.1.30
+
+### Patch Changes
+
+- Updated dependencies [9aad939]
+    - @khanacademy/wonder-blocks-tokens@12.0.1
+    - @khanacademy/wonder-blocks-layout@3.1.30
+
 ## 3.1.29
 
 ### Patch Changes