+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies [9aad939]
+    - @khanacademy/wonder-blocks-tokens@12.0.1
+    - @khanacademy/wonder-blocks-typography@4.2.15
+
 ## 0.3.21
 
 ### Patch Changes