+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.32
+
+### Patch Changes
+
+- Updated dependencies [9aad939]
+    - @khanacademy/wonder-blocks-tokens@12.0.1
+    - @khanacademy/wonder-blocks-typography@4.2.15
+
 ## 5.1.31
 
 ### Patch Changes