+++ packages/wonder-blocks-tooltip/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-tooltip
 
+## 4.1.36
+
+### Patch Changes
+
+- Updated dependencies [9aad939]
+    - @khanacademy/wonder-blocks-tokens@12.0.1
+    - @khanacademy/wonder-blocks-layout@3.1.30
+    - @khanacademy/wonder-blocks-modal@8.1.8
+    - @khanacademy/wonder-blocks-typography@4.2.15
+
 ## 4.1.35
 
 ### Patch Changes